Separate tests files
